pyproject.toml
[tool.poetry]
name = "cfg"
version = "0.1.0"
description = ""
authors = ["Jevgeni Tarassov <jevgeni@tarassov.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
